# dist/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Coppia.co API Documentation",
        "description": "API Documentation for Coppia.co API calls",
        "termsOfService": "http://swagger.io/terms/",
        "contact" : {
            "name" : "Coppia API Support",
            "url" : "www.coppia.co/support",
            "email" : "eric@coppia.co"
        },
        "version": "1.0.0"
    },
    "basePath": "/",
...
    "schemes": [
        "https"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/api/v1/snippets/": {
            "get": {
                "tags": [
                    "Snippets"
                ],
                "description": "Return list of Snippets",
                "operationId": "GetSnippets",
                "produces": [
                    "application/json"
                ],
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Snippet object returned in list",
                        "schema": {
                            "$ref": "#/definitions/snippet"
                        }
                    }
                }
            },
            "post":{
                "tags": [
                    "Snippets"
                ],
                "description": "Creates a new Snippet",
                "operationId": "CreateSnippet",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "text",
                        "in": "body",
                        "description": "Text of the Snippet",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "interview_id",
                        "in": "body",
                        "description": "interview_id for the Snippet",
                        "required": true,
                        "type": "integer",
                        "format": "int32"
                    },
                    {
                        "name": "create_user",
                        "in": "body",
                        "description": "User Id for the logged in user",
                        "required": true,
                        "type": "integer",
                        "format" : "int32"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "When creating a new snippet, the id for the newly created snippet will be returned",
                        "schema": {
                            "$ref": "#/definitions/create_snippet"
                        }
                    }
                }
            }
        },
        "/api/v1/snippets/{snippet_id}": {
            "get":{
                "tags": [
                    "Snippets"
                ],
                "description": "Returns a single Snippet object",
                "operationId": "GetSnippetById",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "snippet_id",
                        "in": "path",
                        "description": "id of the Snippet to return",
                        "required": true,
                        "type": "integer",
                        "format" : "int32"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns single Snippet object",
                        "schema": {
                            "$ref": "#/definitions/snippet"
                        }
                    }
                }
            },
            "put":{
                "tags": [
                    "Snippets"
                ],
                "description": "Returns a success message with count of snippet updates",
                "operationId": "UpdateSnippetById",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "snippet_id",
                        "in": "path",
                        "description": "id for the snippet to update",
                        "required": true,
                        "type": "integer",
                        "format" : "int32"
                    },
                    {
                        "name": "text",
                        "in": "body",
                        "description": "Text of the Snippet",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "interview_id",
                        "in": "body",
                        "description": "interview_id for the Snippet",
                        "required": true,
                        "type": "integer",
                        "format": "int32"
                    },
                    {
                        "name": "update_user",
                        "in": "body",
                        "description": "id for the logged in user",
                        "required": true,
                        "type": "integer",
                        "format" : "int32"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response when calling to update a Snippet",
                        "schema": {
                            "$ref": "#/definitions/update_snippet"
                        }
                    }
                }
            },
            "delete":{
                "tags":[
                    "Snippets"
                ],
                "description":"Deletes a snippet from the database",
                "operationId":"DeleteSnippet",
                "produces": [
                    "application/json"
                ],
                "parameters":[
                    {
                        "name": "snippet_id",
                        "in": "path",
                        "description": "snippet_id of the Snippet to be deleted from the database.",
                        "required": true,
                        "type": "integer",
                        "format" : "int32"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response for the deleted Snippet",
                        "schema":{
                            "$ref":"#/definitions/delete_snippet"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "snippet": {
            "required": [
                "id",
                "text",
                "interview_id",
                "created_by",
                "created_date",
                "updated_by",
                "update_datetime"
            ],
            "properties": {
                "id": {
                    "type": "integer",
                    "format" : "int32",
                    "example": "2"
                },
                "text": {
                    "type": "string",
                    "example": "They are shifting their website to be more mobile friendly. They could see who was using mobile devices. They saw that they needed to make a shift to make the website mobile friendly."
                },
                "interview_id": {
                    "type": "integer",
                    "format": "int32",
                    "example": "2"
                },
                "created_by": {
                    "type": "string",
                    "example": "Eric Hillenbrand"
                },
                "created_date": {
                    "type": "string",
                    "format" : "date-time",
                    "example": "2016-10-17T23:10:23.000Z"
                },
                "created_image_link": {
                    "type": "string",
                    "example": "https://d1ts43dypk8bqh.cloudfront.net/v1/avatars/88a1184e-964b-495b-86fc-92c2867c5427"
                },
                "updated_by": {
                    "type": "integer",
                    "format" : "int32",
                    "example": "Eric Hillenbrand"
                },
                "updated_datetime": {
                    "type": "string",
                    "format" : "date-time",
                    "example": "2016-10-17T23:10:23.000Z"
                },
                "updated_image_link": {
                    "type": "string",
                    "example": "https://d1ts43dypk8bqh.cloudfront.net/v1/avatars/88a1184e-964b-495b-86fc-92c2867c5427"
                }
            }
        },
        "create_snippet": {
            "properties": {
                "snippet_id": {
                    "type": "integer",
                    "format" : "int32",
                    "example" : "3"
                }
            }
        },
         "update_snippet": {
            "properties": {
                "snippet_id": {
                    "type": "integer",
                    "format" : "int32",
                    "example" : "8"
                }
            }
        },
         "delete_snippet": {
            "properties": {
                "snippet_id": {
                    "type": "integer",
                    "format" : "int32",
                    "example" : "6"
                }
            }
        }
    }
}